requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
redis==5.0.7
aiosmtplib==3.0.1
email-validator==2.2.0
httptools==0.6.1
websockets==13.1
